Refactor server into modular package
