+++ shopify_storefront_mcp_server/__init__.py
@@ -1,0 +1,7 @@
+from mcp.server.fastmcp import FastMCP
+
+mcp = FastMCP("shopify_storefront", version="0.3.0")
+
+# Import submodules so that tools/resources are registered when package is loaded
+from . import customer  # noqa: F401
+from . import discovery  # noqa: F401
